pyproject.toml
[project]
name = "microsoft-agent-framework-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.3",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
]
